LayoutTests/http/tests/blink/sendbeacon/beacon-cross-origin.https-expected.txt

CONSOLE MESSAGE: [blocked] The page at https://127.0.0.1:8443/blink/sendbeacon/beacon-cross-origin.https.html was not allowed to display insecure content from http://example.test:8000/blink/sendbeacon/resources/save-beacon.py?name=cross-origin.

CONSOLE MESSAGE: Beacon API cannot load http://example.test:8000/blink/sendbeacon/resources/save-beacon.py?name=cross-origin. Not allowed to request resource
Verify navigator.sendBeacon() mixed content checking.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.sendBeacon("http://example.test:8000/blink/sendbeacon/resources/save-beacon.py?name=cross-origin", "CrossOrigin"); is false
PASS successfullyParsed is true

TEST COMPLETE
